refactor: move document RsapiDao functions to own class
