+++ Gravity/Gravity/DAL/RSAPI/RsapiDao.cs
@@ -1,46 +1,14 @@
-﻿using kCura.Relativity.Client;
-using Relativity.API;
-using Gravity.Globals;
+﻿using System;
 using Gravity.Utils;
+using Relativity.API;
 
 namespace Gravity.DAL.RSAPI
 {
-	using System;
-	using System.Runtime.CompilerServices;
-	using Gravity.Exceptions;
-
-	public partial class RsapiDao
+	public partial class RsapiDao : RsapiDaoBase
 	{
-		public ExecutionIdentity CurrentExecutionIdentity { get; set; }
-
-		protected IHelper helper;
-		protected int workspaceId;
-		
-		private InvokeWithRetryService invokeWithRetryService;
-
-		protected IRSAPIClient CreateProxy()
-		{
-			var proxy = helper.GetServicesManager().CreateProxy<IRSAPIClient>(this.CurrentExecutionIdentity);
-			proxy.APIOptions.WorkspaceID = workspaceId;
-
-			return proxy;
-		}
-
 		public RsapiDao(IHelper helper, int workspaceId, ExecutionIdentity executionIdentity, InvokeWithRetrySettings invokeWithRetrySettings = null)
+			: base(helper, workspaceId, executionIdentity)
 		{
-			this.helper = helper;
-			this.workspaceId = workspaceId;
-			this.CurrentExecutionIdentity = executionIdentity;
-
-			if (invokeWithRetrySettings == null)
-			{
-				InvokeWithRetrySettings defaultSettings = new InvokeWithRetrySettings(SharedConstants.retryAttempts, SharedConstants.sleepTimeInMiliseconds);
-				this.invokeWithRetryService = new InvokeWithRetryService(defaultSettings);
-			}
-			else
-			{
-				this.invokeWithRetryService = new InvokeWithRetryService(invokeWithRetrySettings);
-			}
 		}
 
 		[Obsolete("This constructor has been deprecated. Use RsapiDao(IHelper helper, int workspaceId, ExecutionIdentity executionIdentity, InvokeWithRetrySettings invokeWithRetrySettings) instead.")]
@@ -48,58 +16,5 @@ namespace Gravity.DAL.RSAPI
 			: this(helper, workspaceId, ExecutionIdentity.System, invokeWithRetrySettings)
 		{
 		}
-
-		#region InvokeProxyWithRetry
-		/*
-			These methods (and their overloads) replace the existing retry and error handling logic.
-			When these functions are called, memberName will be replaced by the name of the calling
-			function if not explicitly provided. This is what replaces the MethodInfo.GetCurrentMethod()
-			method when this code was copied to each individual call.
-		*/
-
-		protected T InvokeProxyWithRetry<T>(IRSAPIClient proxy, Func<IRSAPIClient, T> func, [CallerMemberName] string memberName = null)
-		{
-			try
-			{
-				return invokeWithRetryService.InvokeWithRetry(() => func(proxy));
-			}
-			catch (Exception ex)
-			{
-				throw new ProxyOperationFailedException("Failed in method: " + memberName, ex);
-			}
-		}
-
-		protected void InvokeProxyWithRetry(IRSAPIClient proxy, Action<IRSAPIClient> func, [CallerMemberName] string memberName = null)
-		{
-			try
-			{
-				invokeWithRetryService.InvokeVoidMethodWithRetry(() => func(proxy));
-			}
-			catch (Exception ex)
-			{
-				throw new ProxyOperationFailedException("Failed in method: " + memberName, ex);
-			}
-		}
-
-		/*
-			In addition to what the above methods do, these also create the proxy when it is only needed once.
-		*/
-
-		protected T InvokeProxyWithRetry<T>(Func<IRSAPIClient, T> func, [CallerMemberName] string memberName = null)
-		{
-			using (var proxy = CreateProxy())
-			{
-				return InvokeProxyWithRetry(proxy, func, memberName);
-			}
-		}
-
-		protected void InvokeProxyWithRetry(Action<IRSAPIClient> func, [CallerMemberName] string memberName = null)
-		{
-			using (var proxy = CreateProxy())
-			{
-				InvokeProxyWithRetry(proxy, func, memberName);
-			}
-		}
-		#endregion
 	}
 }
