+++ Gravity/Gravity/Gravity.csproj
@@ -60,7 +60,8 @@
     <Compile Include="Attributes\RelativityObjectFieldParentArtifactIdAttribute.cs" />
     <Compile Include="Base\RelativityFile.cs" />
     <Compile Include="Base\AppConfigConnectionHelper.cs" />
-    <Compile Include="DAL\RSAPI\RsapiDao.Get.Document.cs" />
+    <Compile Include="DAL\RSAPI\RsapiDocumentDao.cs" />
+    <Compile Include="DAL\RSAPI\RsapiDaoBase.cs" />
     <Compile Include="Extensions\BaseDtoExtensions.cs" />
     <Compile Include="Extensions\CollectionsExtensions.cs" />
     <Compile Include="Extensions\ResultSetExtensions.cs" />
